feat: restructure handlers by adding new modules for chatgpt, deepl, documents, and translate; remove obsolete utils module
